@@ -51,15 +51,15 @@ func enumerateLines(filename string) chan string {
 	return output
 }
 
-func mapper(filename interface{}, writer mr.Writer, cancel func(error)) {
-	if len(*stopOnFile) > 0 && path.Base(filename.(string)) == *stopOnFile {
+func mapper(filename string, writer mr.Writer[int], cancel func(error)) {
+	if len(*stopOnFile) > 0 && path.Base(filename) == *stopOnFile {
 		fmt.Printf("Stop on file: %s\n", *stopOnFile)
 		cancel(errors.New("stop on file"))
 		return
 	}
 
 	var result int
-	for line := range enumerateLines(filename.(string)) {
+	for line := range enumerateLines(filename) {
 		if strings.HasPrefix(strings.TrimSpace(line), "func") {
 			result++
 		}
@@ -79,11 +79,11 @@ func mapper(filename interface{}, writer mr.Writer, cancel func(error)) {
 	writer.Write(result)
 }
 
-func reducer(input <-chan interface{}, writer mr.Writer, cancel func(error)) {
+func reducer(input <-chan int, writer mr.Writer[int], cancel func(error)) {
 	var result int
 
 	for count := range input {
-		v := count.(int)
+		v := count
 		if *maxFiles > 0 && result >= *maxFiles {
 			fmt.Printf("Reached max files: %d\n", *maxFiles)
 			cancel(errors.New("max files reached"))
@@ -109,7 +109,7 @@ func main() {
 	fmt.Println("Processing, please wait...")
 
 	start := time.Now()
-	result, err := mr.MapReduce(func(source chan<- interface{}) {
+	result, err := mr.MapReduce(func(source chan<- string) {
 		filepath.Walk(*dir, func(fpath string, f os.FileInfo, err error) error {
 			if !f.IsDir() && path.Ext(fpath) == ".go" {
 				source <- fpath
