@@ -8,19 +8,19 @@ import (
 )
 
 func main() {
-	mr.MapReduceVoid(func(source chan<- interface{}) {
+	mr.MapReduceVoid(func(source chan<- int) {
 		for i := 0; i < 10; i++ {
 			source <- i
 		}
-	}, func(item interface{}, writer mr.Writer, cancel func(error)) {
-		i := item.(int)
+	}, func(item int, writer mr.Writer[int], cancel func(error)) {
+		i := item
 		if i == 0 {
 			time.Sleep(10 * time.Second)
 		} else {
 			time.Sleep(5 * time.Second)
 		}
 		writer.Write(i)
-	}, func(pipe <-chan interface{}, cancel func(error)) {
+	}, func(pipe <-chan int, cancel func(error)) {
 		for i := range pipe {
 			fmt.Println(i)
 		}
